STAR-515 Add guardrail for SASI and STAR-512 Add guardrails for total and per-table indexes for SAI (#140)
* STAR-512 Add guardrails for total and per-table indexes for SAI


* STAR-515 Add guardrail for SASI


(cherry picked from commit e2ae18ead10d7186a76bb1a8299bfcd4ddf3820e)
(cherry picked from commit eaf8c59b3820a575742370aa6b4936d3783bc49b)
(cherry picked from commit 753fb6916ba32673a479b8c8e39cec447aa97012)
(cherry picked from commit 0fafd7b053fc2520219ab8a0243362539696baf8)
(cherry picked from commit 41d26d3e2dc86f8be5e1639ac68d684eb3e57772)
(cherry picked from commit 3a40eae385b31cac03b61cb81fd5042c0f93964c)
(cherry picked from commit 203e112c8ac22fb9141913b15e3b9d9b5b36f30a)
(cherry picked from commit d508237f522f941c33c00c5ad97beb9d7ab09aab)

+++ test/unit/org/apache/cassandra/guardrails/GuardrailSecondaryIndexesPerTableTest.java
@@ -29,13 +29,13 @@ import static org.junit.Assert.assertEquals;
 
 public class GuardrailSecondaryIndexesPerTableTest extends GuardrailTester
 {
-    private Long defaultSIPerTableFailureThreshold;
+    private int defaultSIPerTableFailureThreshold;
 
     @Before
     public void before()
     {
         defaultSIPerTableFailureThreshold = config().secondary_index_per_table_failure_threshold;
-        config().secondary_index_per_table_failure_threshold = 1L;
+        config().secondary_index_per_table_failure_threshold = 1;
     }
 
     @After
